Resume operation on broken database entries
When backing up his phone a user noticed that some entries may be broken
and could not be decoded by Firefox. If that happens resume the
operation with the next entry. Unfortunately that means reopening
everything again, as Firefox seems to loose its complete context after
such an error, so several global variables are required to track the
state.

Thanks a lot to Philipp for testing dozens of proposed fixes as I wasn't
able to reproduce this error myself for a long time!